StadiumAiChatbot/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
psycopg2
langdetect